ecommerce/tests/factories.py
import factory
from products.models import Product, Category
from users.models import UserProfile
from django.contrib.auth.models import User

class CategoryFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Category
    name = factory.Faker("word")

class ProductFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = Product
    name = factory.Faker("word")
    price = factory.Faker("random_int", min=1, max=100)
    category = factory.SubFactory(CategoryFactory)

class UserFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = User
    username = factory.Faker("user_name")
    email = factory.Faker("email")

class UserProfileFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = UserProfile
    user = factory.SubFactory(UserFactory)
